requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database and Supabase
supabase==2.0.3
psycopg2-binary==2.9.9

# Environment and Configuration
python-dotenv==1.0.0

# Data Processing and ML
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# Image Processing
Pillow==10.1.0

# Authentication
PyJWT==2.8.0

# Optional: For Jupyter Notebooks (uncomment if using notebooks)
# jupyter==1.0.0
# matplotlib==3.8.2
# seaborn==0.13.0
# ipykernel==6.27.1

# Development Tools (optional, uncomment for development)
# ipython==8.18.1
# pytest==7.4.3
# pytest-django==4.7.0


